AI_VideoStream/Face_Detection/requirements.txt
deepface==0.0.75
Flask==2.2.2
numpy==1.19.5
opencv-python==4.6.0.66
pandas==1.4.4
pillow==8.4.0
retina-face==0.0.12

